# yaml-first/src/main/java/org/os890/sketch/petstore/PetResource.java
/*
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package org.os890.sketch.petstore;

import java.time.OffsetDateTime;
import java.util.List;
import java.util.Map;
import java.util.concurrent.ConcurrentHashMap;
import java.util.concurrent.atomic.AtomicLong;

import org.os890.sketch.petstore.dto.ApiError;
import org.os890.sketch.petstore.dto.NewPet;
import org.os890.sketch.petstore.dto.Pet;
import org.os890.sketch.petstore.dto.PetStatus;
import jakarta.validation.Valid;
import jakarta.ws.rs.Consumes;
import jakarta.ws.rs.GET;
import jakarta.ws.rs.POST;
import jakarta.ws.rs.Path;
import jakarta.ws.rs.PathParam;
import jakarta.ws.rs.Produces;
import jakarta.ws.rs.QueryParam;
import jakarta.ws.rs.core.MediaType;
import jakarta.ws.rs.core.Response;

/**
 * Hand-written JAX-RS 3.1 resource demonstrating usage of the DTO classes
 * generated by the openapi-generator-maven-plugin (see target/generated-sources/openapi).
 */
@Path("/pets")
@Produces(MediaType.APPLICATION_JSON)
@Consumes(MediaType.APPLICATION_JSON)
public class PetResource {

    private final Map<Long, Pet> pets = new ConcurrentHashMap<>();
    private final AtomicLong idSequence = new AtomicLong(1);

    @GET
    public List<Pet> listPets(@QueryParam("status") PetStatus status) {
        return pets.values().stream()
                .filter(pet -> status == null || pet.getStatus() == status)
                .toList();
    }

    @GET
    @Path("/{petId}")
    public Response getPet(@PathParam("petId") long petId) {
        Pet pet = pets.get(petId);
        if (pet == null) {
            ApiError error = new ApiError()
                    .code(404)
                    .message("Pet " + petId + " not found");
            return Response.status(Response.Status.NOT_FOUND).entity(error).build();
        }
        return Response.ok(pet).build();
    }

    @POST
    public Response createPet(@Valid NewPet newPet) {
        Pet pet = new Pet()
                .id(idSequence.getAndIncrement())
                .name(newPet.getName())
                .status(newPet.getStatus() != null ? newPet.getStatus() : PetStatus.AVAILABLE)
                .category(newPet.getCategory())
                .tags(newPet.getTags())
                .createdAt(OffsetDateTime.now());
        pets.put(pet.getId(), pet);
        return Response.status(Response.Status.CREATED).entity(pet).build();
    }
}